fixing errors from the front 2

+++ src/main/java/com/locadora/LocadoradeLivrosApi/service/AlugueisService.java
@@ -1,10 +1,11 @@
 package com.locadora.LocadoradeLivrosApi.service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
+
 
 import org.springframework.stereotype.Service;
 
@@ -38,10 +39,18 @@ public class AlugueisService {
 		
 		Livro livro  = livroRepository.getById(aluguel.getLivro().getId());
 		
+		LocalDate dataAtual = LocalDate.now();
+		
+		aluguel.setData_devolucao(null);
+		
 		if(aluguel.getData_previsao().isBefore(aluguel.getData_aluguel())) {
 			throw new com.locadora.LocadoradeLivrosApi.service.exceptions.DataIntegrityViolationException(
 					"A data de previsão não pode ser maior que a de aluguel");
 		}	
+		else if(aluguel.getData_previsao().isBefore(aluguel.getData_aluguel())) {
+			throw new com.locadora.LocadoradeLivrosApi.service.exceptions.DataIntegrityViolationException(
+					"A data de aluguel deve ser a data atual: "+ dataAtual);
+		}
 		else {				
 		
 		if(livro.getQuantidade() > 0) {
@@ -65,10 +74,17 @@ public class AlugueisService {
 		
 		Livro livro  = livroRepository.findById(aluguel.getLivro().getId()).orElseThrow(null);
 		
+		LocalDate dataAtual = LocalDate.now();
+		
 		if(aluguel.getData_devolucao().isBefore(aluguel.getData_aluguel())) {
 			throw new com.locadora.LocadoradeLivrosApi.service.exceptions.DataIntegrityViolationException(
 					"A data de devolução não pode ser maior que a de aluguel");
-		}	
+		
+		}else if(aluguel.getData_devolucao().isAfter(dataAtual) || aluguel.getData_devolucao().isBefore(dataAtual)){
+			throw new com.locadora.LocadoradeLivrosApi.service.exceptions.DataIntegrityViolationException(
+					"A data de devolução deve ser a data atual: "+ dataAtual);
+		
+		}
 		else {
 		if(aluguel.getData_devolucao() != null) {
 			livro.setQuantidade(livro.getQuantidade() + 1);
@@ -84,11 +100,15 @@ public class AlugueisService {
 	}
 
 	public void deletarAluguel(Aluguel aluguel) {
-		try {
-			aluguelRepository.delete(aluguel);
-		} catch (DataIntegrityViolationException e) {
+		Livro livro  = livroRepository.findById(aluguel.getLivro().getId()).orElseThrow(null);
+		if(aluguel.getData_devolucao() != null) {
 			throw new com.locadora.LocadoradeLivrosApi.service.exceptions.DataIntegrityViolationException(
 					"Este aluguel não pode ser deletado, possui livros e usuários cadastrados");
 		}
+		else {
+			livro.setQuantidade(livro.getQuantidade() + 1);
+			aluguelRepository.delete(aluguel);
+		}
+		
 	}
 }
